backend/pyproject.toml
[project]
name = "synergyx-backend"
version = "0.1.0"
description = "SynergyX-AI Backend — Autonomous team-building platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlmodel>=0.0.22",
    "openai>=1.60.0",
    "pydantic>=2.10.0",
    "pypdf2>=3.0.0",
    "python-multipart>=0.0.18",
    "python-dotenv>=1.0.0",
    "numpy>=2.0.0",
]
